requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
openai>=1.29.0
google-cloud-aiplatform>=1.30.0
python-dotenv>=1.0.0
httpx>=0.24.0
pydantic>=1.10.7
tenacity>=8.2.2
